refact: #1 PostsListTitle Component로 분리
